# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      tags: [Staff]
      operationId: listStaff
      summary: List all staff users (including invited-only allowlist entries)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of staff.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffMember'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Staff]
      operationId: inviteStaff
      summary: Add a staff member to the allowlist (admin only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteStaffBody'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMember'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /staff/{id}:
    patch:
      tags: [Staff]
      operationId: updateStaff
      summary: Update a staff member's role or active status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffBody'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffMember'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Staff]
      operationId: removeStaff
      summary: Remove a staff member from the allowlist and system (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '204':
          description: Removed.
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /trucks:
    get:
      tags: [Trucks]
      operationId: listTrucks
      summary: List all trucks
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of trucks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
    post:
      tags: [Trucks]
      operationId: createTruck
      summary: Create a truck
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTruckBody'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
  /trucks/{id}:
    patch:
      tags: [Trucks]
      operationId: updateTruck
      summary: Update a truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTruckBody'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    StaffMember:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - profileImageUrl
        - role
        - active
        - status
        - lastLoginAt
        - createdAt
      properties:
        id: { type: string }
        email: { type: [string, 'null'], format: email }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        profileImageUrl: { type: [string, 'null'] }
        role:
          type: string
          enum: [admin, staff, warehouse, driver]
        active: { type: boolean }
        status:
          type: string
          enum: [invited, active, inactive]
        lastLoginAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    UpdateStaffBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, staff, warehouse, driver]
        active: { type: boolean }
    InviteStaffBody:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          minLength: 1
        role:
          type: string
          enum: [admin, staff, warehouse, driver]
...
    Truck:
      type: object
      required: [id, name, active, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateTruckBody:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    UpdateTruckBody:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        active: { type: boolean }
...